Split local.rs into originate and deliver.rs for clarity
