+++ bpa/docs/TODO.md
@@ -147,6 +147,6 @@ Open bpa-side items from the `refactor/bpv7-parse` deep review (`references/revi
 
 **Egress filters see pre-increment cached fields (B5 residual).** `forward.rs` replaces only `bundle.bpv7.blocks` after `update_extension_blocks`, so the Egress filter chain reads pre-increment `metadata.extensions.{previous_node, hop_count, age}`. The wire is correct and the filter-chain path refreshes after write-filter rewrites, but an egress filter enforcing a hop-count limit on this node reads the stale count — and the justifying comment ("never observed again after this rewrite") overlooks that egress filters do observe `metadata.extensions`. Either refresh the cached fields before the filter chain runs or fix the comment and document the filter-visible staleness as contract.
 
-**Hot-path structural re-parses (B8).** Four hot paths re-run `hardy_bpv7::parse::parse` on stored bytes the BPA has already parsed: `forward.rs` per forward *attempt* (repeated on every retry/requeue), `local.rs` per application delivery (to recover `bcb_ops` for payload decryption), `admin.rs` per admin record, and `adu_reassembly.rs` per fragment merge. The structural-bundle restructure kept the re-parses because OperationSets are not persisted. Candidates: persist/carry the OperationSets in metadata, or cache the parse alongside the loaded data. Natural home: the filters phase or a dedicated perf tranche; benchmark before and after. Sibling inside the bpv7 BPSec cascade (parse-review finding 5 / refactor-parse review R-07): `canonical_primary` re-decodes and CRC-re-verifies the primary block on every include-primary BIB/BCB operation, pure confirmation of an invariant the parser already enforced — pass an "already canonical" flag or cache the canonical primary bytes once per bundle so the IPPT/AAD path borrows raw.
+**Hot-path structural re-parses (B8).** Four hot paths re-run `hardy_bpv7::parse::parse` on stored bytes the BPA has already parsed: `forward.rs` per forward *attempt* (repeated on every retry/requeue), `deliver.rs` per application delivery (to recover `bcb_ops` for payload decryption), `admin.rs` per admin record, and `adu_reassembly.rs` per fragment merge. The structural-bundle restructure kept the re-parses because OperationSets are not persisted. Candidates: persist/carry the OperationSets in metadata, or cache the parse alongside the loaded data. Natural home: the filters phase or a dedicated perf tranche; benchmark before and after. Sibling inside the bpv7 BPSec cascade (parse-review finding 5 / refactor-parse review R-07): `canonical_primary` re-decodes and CRC-re-verifies the primary block on every include-primary BIB/BCB operation, pure confirmation of an invariant the parser already enforced — pass an "already canonical" flag or cache the canonical primary bytes once per bundle so the IPPT/AAD path borrows raw.
 
 **De-flake `storage::channel::tests::test_hysteresis_recovery`.** Load-dependent timing flake (passes in isolation, fails occasionally under the full parallel run). Pre-existing hysteresis-timing sensitivity, untouched by the train; fix the test's timing assumptions rather than loosening the hysteresis.
