+++ bpa/src/dispatcher/mod.rs
@@ -4,10 +4,11 @@ use hardy_bpv7::{eid::Eid, status_report::ReasonCode};
 use super::*;
 
 mod admin;
+mod deliver;
 mod dispatch;
 mod forward;
 mod ingress;
-mod local;
+mod originate;
 mod reassemble;
 mod report;
 mod restart;
